+++ resources/views/admin/includes/common_js/bulk_action_js.blade.php
@@ -3,7 +3,9 @@ $("#bulk_action").on("click",function(){
     let table = $('#dataListTable').DataTable();
     idsArray = table.rows({selected: true}).ids().toArray();
     let route_name_bulk_active_inactive = "{{route($route_name_bulk_active_inactive)}}";
-    let route_name_bulk_delete = "{{route($route_name_bulk_delete)}}";
+    @if (isset($route_name_bulk_delete))
+        let route_name_bulk_delete = "{{route($route_name_bulk_delete)}}";
+    @endif
 
 
     if(idsArray.length === 0){
@@ -30,10 +32,10 @@ $("#bulk_action").on("click",function(){
                             $('#alert_message').fadeOut("slow");
                         }, 3000);
                     }
-                    else if (data.disabled_demo) {
+                    else if (data.demo) {
                         $('#bulkConfirmModal').modal('hide');
                         $('#alert_message').fadeIn("slow"); //Check in top in this blade
-                        $('#alert_message').addClass('alert alert-danger').html(data.disabled_demo);
+                        $('#alert_message').addClass('alert alert-danger').html(data.demo);
                         setTimeout(function() {
                             $('#alert_message').fadeOut("slow");
                         }, 3000);
@@ -60,10 +62,10 @@ $("#bulk_action").on("click",function(){
                             $('#alert_message').fadeOut("slow");
                         }, 3000);
                     }
-                    else if (data.disabled_demo) {
+                    else if (data.demo) {
                         $('#bulkConfirmModal').modal('hide');
                         $('#alert_message').fadeIn("slow"); //Check in top in this blade
-                        $('#alert_message').addClass('alert alert-danger').html(data.disabled_demo);
+                        $('#alert_message').addClass('alert alert-danger').html(data.demo);
                         setTimeout(function() {
                             $('#alert_message').fadeOut("slow");
                         }, 3000);
@@ -72,35 +74,37 @@ $("#bulk_action").on("click",function(){
             });
         });
 
-        //Bulk Delete
-        $("#bulkDelete").on("click",function(){
-            $.ajax({
-                url: route_name_bulk_delete,
-                method: "GET",
-                data: {idsArray:idsArray},
-                success: function (data) {
-                    console.log(data);
-                    if(data.success){
-                        $('#bulkConfirmModal').modal('hide');
-                        table.rows('.selected').deselect();
-                        $('#dataListTable').DataTable().ajax.reload();
-                        $('#alert_message').fadeIn("slow"); //Check in top in this blade
-                        $('#alert_message').addClass('alert alert-success').html(data.success);
-                        setTimeout(function() {
-                            $('#alert_message').fadeOut("slow");
-                        }, 3000);
+        @if (isset($route_name_bulk_delete))
+            $("#bulkDelete").on("click",function(){
+                $.ajax({
+                    url: "{{route($route_name_bulk_delete)}}",
+                    method: "GET",
+                    data: {idsArray:idsArray},
+                    success: function (data) {
+                        console.log(data);
+                        if(data.success){
+                            $('#bulkConfirmModal').modal('hide');
+                            table.rows('.selected').deselect();
+                            $('#dataListTable').DataTable().ajax.reload();
+                            $('#alert_message').fadeIn("slow"); //Check in top in this blade
+                            $('#alert_message').addClass('alert alert-success').html(data.success);
+                            setTimeout(function() {
+                                $('#alert_message').fadeOut("slow");
+                            }, 3000);
+                        }
+                        else if (data.demo) {
+                            $('#bulkConfirmModal').modal('hide');
+                            $('#dataListTable').DataTable().ajax.reload();
+                            $('#alert_message').fadeIn("slow"); //Check in top in this blade
+                            $('#alert_message').addClass('alert alert-danger').html(data.demo);
+                            setTimeout(function() {
+                                $('#alert_message').fadeOut("slow");
+                            }, 3000);
+                        }
                     }
-                    else if (data.disabled_demo) {
-                        $('#bulkConfirmModal').modal('hide');
-                        $('#dataListTable').DataTable().ajax.reload();
-                        $('#alert_message').fadeIn("slow"); //Check in top in this blade
-                        $('#alert_message').addClass('alert alert-danger').html(data.disabled_demo);
-                        setTimeout(function() {
-                            $('#alert_message').fadeOut("slow");
-                        }, 3000);
-                    }
-                }
+                });
             });
-        });
+        @endif
+
     }
 });
